+++ src/library/Telegram.BotAPI/AvailableTypes/Voice.cs
@@ -37,5 +37,5 @@ public class Voice
     /// Optional. File size in bytes. It can be bigger than 2^31 and some programming languages may have difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type are safe for storing this value.
     /// </summary>
     [JsonPropertyName(PropertyNames.FileSize)]
-    public int? FileSize { get; set; }
+    public long? FileSize { get; set; }
 }
